Make query in startup
